Step 4.0
wip step 4
