+++ ocr/src/main/java/com/ing/software/ocr/Legacy/OcrUtilsDep.java
@@ -1,0 +1,8 @@
+package com.ing.software.ocr.Legacy;
+
+/**
+ *
+ */
+@Deprecated
+public class OcrUtilsDep {
+}
