+++ ocr/src/main/java/com/ing/software/ocr/OcrObjects/OcrOptions.java
@@ -5,11 +5,21 @@ import android.support.annotation.IntRange;
 /**
  * Object passed to the manager to avoid performing unnecessary operations
  * and consequently reduce time (time depends primarily on precision)
+ *
+ * For precision:
+ * 0 = scan image at 1/3 of its dimension, don't reanalyze specific parts of image to get better results
+ * 1 = scan image at 1/2 of its dimension, don't reanalyze specific parts of image to get better results
+ * 2 = scan image at original dimension (passed by imageprocessor), don't reanalyze specific parts of image to get better results
+ * 3 = scan image at original dimension (passed by imageprocessor), reanalyze total strip to get a better result
+ * 4 = scan image at original dimension (passed by imageprocessor), reanalyze total and prices strips to get a better result
+ * 5 = scan image at original dimension (passed by imageprocessor), reanalyze total and prices strips to get a better result, if
+ *      nothing was found scan also upside down
  */
 
 public class OcrOptions {
 
-    private static final int DEFAULT_PRECISION = 2;
+    public static final int REDO_OCR_PRECISION = 3;
+    private static final int DEFAULT_PRECISION = 3;
     private static final boolean DEFAULT_TOTAL = true;
     private static final boolean DEFAULT_DATE = true;
     private static final boolean DEFAULT_PRODUCTS = true;
