+++ ocr/src/main/java/com/ing/software/ocr/OcrVars.java
@@ -24,11 +24,12 @@ public class OcrVars {
     public static final String PRODUCTS_TAG = "products"; //tag for rawtext on central-left part of the receipt
     public static final String PRICES_TAG = "prices"; //tag for rawtext on central-right part of the receipt
     public static final String CONCLUSION_TAG = "conclusion"; //tag for rawtext on bottom of the receipt
-    static final double NUMBER_MIN_VALUE = 0.4; //Max allowed value to accept a string as a valid number. See OcrUtils.isPossiblePriceNumber()
-    static final double NUMBER_MIN_VALUE_ALTERNATIVE = 0.1; //Max allowed value to accept a string as a valid number. See OcrUtils.isPossiblePriceNumber(). Used in alternative search
-    static final int HEIGHT_CENTER_DIFF_MULTIPLIER = 50; //Multiplier used while analyzing difference in alignment between the center of two rects in dataAnalyzer
+    public static final double NUMBER_MIN_VALUE = 0.4; //Max allowed value to accept a string as a valid number. See OcrUtils.isPossiblePriceNumber()
+    public static final double NUMBER_MIN_VALUE_ALTERNATIVE = 0.1; //Max allowed value to accept a string as a valid number. See OcrUtils.isPossiblePriceNumber(). Used in alternative search
+    public static final int HEIGHT_CENTER_DIFF_MULTIPLIER = 50; //Multiplier used while analyzing difference in alignment between the center of two rects in dataAnalyzer
     public static final int HEIGHT_LIST_MULTIPLIER = 80; //Multiplier used while analyzing difference between average height of rects and a specific rect. Used in ProbGrid.getRectHeightScore()
     public static final int HEIGHT_SOURCE_DIFF_MULTIPLIER = 50; //Multiplier used while analyzing difference in height between source and target rect (total with it's price)
+    public static final float RECT_HEIGHT_EXTENDER = 1.5f; //Extend height of chosen rect
 
 
     // day 1 to 31 with or without 0 tens.
@@ -63,7 +64,7 @@ public class OcrVars {
     //In principle, multiple words should be matched with a space between them,
     //but since sometimes some words are split into multiple words, I remove all spaces all together
     //and match the words without spaces, even if there were in origin effectively distinct words.
-    static final List<WordMatcher> AMOUNT_MATCHERS = asList(
+    public static final List<WordMatcher> AMOUNT_MATCHERS = asList(
             new WordMatcher("T[OUD]TALE", 1),
             new WordMatcher("TOT", 0),
             new WordMatcher("T[OUD]TALEE[UI]R[OD]", 3),
@@ -74,9 +75,6 @@ public class OcrVars {
             new WordMatcher("CONTANT[EI]", 1),
             new WordMatcher("CARTADICREDITO", 3),
             new WordMatcher("PAGAMENTOCONTANTE", 4)
-            //new WordMatcher("CCRED", 0) ?
-            //new WordMatcher("ASSEGNI", 1) ?
-            //new WordMatcher("ARROTOND", 0) ?
     );
     static final List<WordMatcher> CHANGE_MATCHERS = asList(
             new WordMatcher("RESTO", 1)
@@ -84,8 +82,4 @@ public class OcrVars {
     static final List<WordMatcher> INDOOR_MATCHERS = asList(
             new WordMatcher("COPERTO", 1)
     );
-
-    // ideal character width / height
-    static final double CHAR_ASPECT_RATIO = 5./8.;
-
 }
