+++ ocr/src/main/java/com/ing/software/ocr/OperativeObjects/ListAmountOrganizer.java
@@ -1,5 +1,7 @@
 package com.ing.software.ocr.OperativeObjects;
 
+import android.support.annotation.NonNull;
+
 import com.ing.software.common.Scored;
 import com.ing.software.ocr.OcrObjects.TempText;
 import com.ing.software.ocr.ScoreFunc;
@@ -11,16 +13,16 @@ import java.util.List;
  * keep ordered list of possible results from target string
  */
 
-public class ListOrganizer {
+public class ListAmountOrganizer implements Comparable<ListAmountOrganizer>{
 
     private Scored<TempText> sourceText;
     private List<Scored<TempText>> targetTexts = new ArrayList<>();
 
-    public ListOrganizer(TempText source) {
-        sourceText = new Scored<>(ScoreFunc.getSourceAmountScore(source), source);
+    public ListAmountOrganizer(Scored<TempText> source) {
+        sourceText = new Scored<>(ScoreFunc.getSourceAmountScore(source), source.obj());
     }
 
-    public void setTargetTexts(List<TempText> targets) {
+    public void setAmountTargetTexts(List<TempText> targets) {
         for (TempText text : targets) {
             targetTexts.add(new Scored<>(ScoreFunc.getAmountScore(text), text));
         }
@@ -32,5 +34,10 @@ public class ListOrganizer {
 
     public List<Scored<TempText>> getTargetTexts() {
         return targetTexts;
+    }
+
+    @Override
+    public int compareTo(@NonNull ListAmountOrganizer listAmountOrganizer) {
+        return sourceText.compareTo(listAmountOrganizer.getSourceText());
     }
 }
